Rename random -> utils
