+++ packages/common/src/utils/arrays/index.ts
@@ -1,2 +1,2 @@
-export * from './random';
+export * from './utils';
 export * from './reducers';
